requirements.txt
artifactsmmo_wrapper==3000.1.1
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10
requests==2.32.3
urllib3==2.4.0
python-dotenv==1.0.0
